+++ components/GoalInput.tsx
@@ -1,0 +1,49 @@
+import { useState } from "react";
+import { Button, StyleSheet, TextInput, View } from "react-native";
+
+function GoalInput(props) {
+  const [enteredGoalText, setEnteredGoalText] = useState("");
+
+  function goalInputHandler(enteredText) {
+    setEnteredGoalText(enteredText);
+  }
+
+  function addGoalHandler() {
+    props.onPressAddGoal(enteredGoalText);
+    setEnteredGoalText("");
+  }
+
+  return (
+    <View style={styles.inputContainer}>
+      <TextInput
+        style={styles.textInput}
+        placeholder="input your goal here"
+        onChangeText={goalInputHandler}
+        value={enteredGoalText}
+      />
+      <Button title="Add Goal" onPress={addGoalHandler} />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  inputContainer: {
+    flexDirection: "row",
+    flex: 1,
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: 24,
+    borderBottomWidth: 1,
+    borderBottomColor: "#CCCCCC",
+  },
+  textInput: {
+    borderWidth: 1,
+    borderColor: "#808080",
+    width: "70%",
+    marginRight: 8,
+    padding: 8,
+  },
+});
+
+export default GoalInput;
+
