+++ references/__tests__/ThemedText-test.tsx
@@ -1,10 +1,11 @@
-import * as React from 'react';
-import renderer from 'react-test-renderer';
+import * as React from "react";
+import renderer from "react-test-renderer";
 
-import { ThemedText } from '../ThemedText';
+import { ThemedText } from "../ThemedText";
 
 it(`renders correctly`, () => {
   const tree = renderer.create(<ThemedText>Snapshot test!</ThemedText>).toJSON();
 
   expect(tree).toMatchSnapshot();
 });
+
